added wpMapping
